Added Cached Storage to RDBStorage
