+++ optuna/storages/__init__.py
@@ -2,7 +2,6 @@ from typing import Union
 
 from optuna._callbacks import RetryFailedTrialCallback
 from optuna.storages._base import BaseStorage
-from optuna.storages._cached_storage import _CachedStorage
 from optuna.storages._heartbeat import fail_stale_trials
 from optuna.storages._in_memory import InMemoryStorage
 from optuna.storages._journal.base import BaseJournalLogStorage
@@ -13,6 +12,8 @@ from optuna.storages._journal.redis import JournalRedisStorage
 from optuna.storages._journal.storage import JournalStorage
 from optuna.storages._rdb.storage import RDBStorage
 
+
+_CachedStorage = RDBStorage._CachedStorage
 
 __all__ = [
     "BaseStorage",
